requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
passlib
bcrypt
python-jose
python-multipart
python-dotenv
pydantic[email]
firebase_admin
fastapi-mail
psycopg2-binary
django
dj_database_url
whitenoise
pillow
django-cors-headers
strawberry-graphql[fastapi]